refactor(src, test) validate 폴더 위치 수정
