+++ src/view/input/validate/BridgeLengthException.js
@@ -1,5 +1,5 @@
 const Exceptionable = require('./Exceptionable');
-const { ERROR_MESSAGE, REGEX } = require('../../utils/constants');
+const { ERROR_MESSAGE, REGEX } = require('../../../utils/constants');
 
 class BrdgeLengthException extends Exceptionable {
   #bridgeLength;
