+++ src/view/input/validate/Exceptionable.js
@@ -1,6 +1,5 @@
 class Exceptionable {
-  #OVERIDING_ERROR_MESSAGE =
-    '[ERROR] YOU SHOULD DECLARE "getValidation method"';
+  #OVERIDING_ERROR_MESSAGE = '[ERROR] YOU SHOULD DECLARE "getValidation method"';
 
   isValidate() {
     throw new Error(this.#OVERIDING_ERROR_MESSAGE);
